Call killpg only if setsid is enabled

+++ waflib/Utils.py
@@ -865,9 +865,9 @@ def run_regular_process(cmd, kwargs, cargs={}):
 		try:
 			out, err = proc.communicate(**cargs)
 		except TimeoutExpired:
-			try:
+			if kwargs.get('start_new_session') and hasattr(os, 'killpg'):
 				os.killpg(proc.pid, signal.SIGKILL)
-			except AttributeError:
+			else:
 				proc.kill()
 			out, err = proc.communicate()
 			raise TimeoutExpired(proc.args, timeout=cargs['timeout'], output=out, stderr=err)
@@ -877,9 +877,9 @@ def run_regular_process(cmd, kwargs, cargs={}):
 		try:
 			status = proc.wait(**cargs)
 		except TimeoutExpired as e:
-			try:
+			if kwargs.get('start_new_session') and hasattr(os, 'killpg'):
 				os.killpg(proc.pid, signal.SIGKILL)
-			except AttributeError:
+			else:
 				proc.kill()
 			proc.wait()
 			raise e
